Web:Moved the FloatingButton component from the Common module to the Components module.
