+++ packages/client/src/components/panels/VersionHistoryPanel/index.js
@@ -4,7 +4,7 @@ import Backdrop from "@docspace/components/backdrop";
 import Heading from "@docspace/components/heading";
 import Aside from "@docspace/components/aside";
 import Loaders from "@docspace/common/components/Loaders";
-import FloatingButton from "@docspace/common/components/FloatingButton";
+import FloatingButton from "@docspace/components/FloatingButton";
 import { withTranslation } from "react-i18next";
 import {
   StyledVersionHistoryPanel,
