+++ packages/client/src/pages/PortalSettings/categories/data-management/backup/auto-backup/index.js
@@ -27,15 +27,12 @@ import ThirdPartyStorageModule from "./sub-components/ThirdPartyStorageModule";
 //import { getThirdPartyCommonFolderTree } from "@docspace/common/api/files";
 import ButtonContainer from "./sub-components/ButtonContainer";
 import AutoBackupLoader from "@docspace/common/components/Loaders/AutoBackupLoader";
-import FloatingButton from "@docspace/common/components/FloatingButton";
+import FloatingButton from "@docspace/components/FloatingButton";
 import Badge from "@docspace/components/badge";
 import { getSettingsThirdParty } from "@docspace/common/api/files";
 
-const {
-  DocumentModuleType,
-  ResourcesModuleType,
-  StorageModuleType,
-} = BackupStorageType;
+const { DocumentModuleType, ResourcesModuleType, StorageModuleType } =
+  BackupStorageType;
 const { EveryDayType, EveryWeekType, EveryMonthType } = AutoBackupPeriod;
 class AutomaticBackup extends React.PureComponent {
   constructor(props) {
@@ -591,10 +588,8 @@ export default inject(
       defaultFolderId,
     } = backup;
 
-    const {
-      updateBaseFolderPath,
-      resetNewFolderPath,
-    } = selectFolderDialogStore;
+    const { updateBaseFolderPath, resetNewFolderPath } =
+      selectFolderDialogStore;
 
     const isCheckedDocuments = selectedStorageType === `${DocumentModuleType}`;
     const isCheckedThirdParty =
