+++ packages/common/components/Section/index.js
@@ -20,7 +20,7 @@ import SubInfoPanelBody from "./sub-components/info-panel-body";
 import SubInfoPanelHeader from "./sub-components/info-panel-header";
 import SubSectionFooter from "./sub-components/section-footer";
 
-import FloatingButton from "../FloatingButton";
+import FloatingButton from "@docspace/components/FloatingButton";
 
 const Section = (props) => {
   const {
