+++ packages/components/FloatingButton/FloatingButton.stories.js
